+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for lovely.esdb
 unreleased
 ==========
 
+ - added relation properties for 1:1 and 1:n relations
+
 2016/08/10 0.3.1
 ================